Implementation de murs multiples

+++ src/Cosinus.cpp
@@ -13,7 +13,7 @@ double Cosinus::cos(double x) const
 	x = std::fmod(x, 2.0 * M_PI);
 	if (x < 0)
 		x += 2.0 * M_PI;
-    int index = (double)_values.size() * x / (2.0 * M_PI);
+    int index = (int)((double)_values.size() * x / (2.0 * M_PI));
 	return _values[index];
 	return 0;
 }
